Remove Rules.csproj (#5218)
