+++ SonarLint.VisualStudio.Integration.sln
@@ -148,12 +148,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessJarFiles", "build\Pr
 		{E47E649E-5044-485B-A077-8FB282094A7D} = {E47E649E-5044-485B-A077-8FB282094A7D}
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Rules", "Rules", "{EFA34F81-1751-49D7-8F62-4085E75FA5F2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rules", "src\Rules\Rules.csproj", "{BA48309C-FB66-47D3-997D-D3068C4F719B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rules.UnitTests", "src\Rules.UnitTests\Rules.UnitTests.csproj", "{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_tools", "_tools", "{AE914B1E-AFE9-4CC4-8E40-B2FB8AD392D9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExtractRuleDescFromJson", "build\tools\ExtractRuleDescFromJson\ExtractRuleDescFromJson.csproj", "{E47E649E-5044-485B-A077-8FB282094A7D}"
@@ -174,9 +168,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SLCore", "src\SLCore\SLCore
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SLCore.UnitTests", "src\SLCore.UnitTests\SLCore.UnitTests.csproj", "{72C861AC-2CAB-46C0-9A9E-06096F589395}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SLCore.Listeners", "src\SLCore.Listeners\SLCore.Listeners.csproj", "{8523D19D-1760-4608-9A4F-3A9353279A16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SLCore.Listeners", "src\SLCore.Listeners\SLCore.Listeners.csproj", "{8523D19D-1760-4608-9A4F-3A9353279A16}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SLCore.Listeners.UnitTests", "src\SLCore.Listeners.UnitTests\SLCore.Listeners.UnitTests.csproj", "{23321F44-DE37-4801-96BB-BC8E3CCFC7F1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SLCore.Listeners.UnitTests", "src\SLCore.Listeners.UnitTests\SLCore.Listeners.UnitTests.csproj", "{23321F44-DE37-4801-96BB-BC8E3CCFC7F1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -466,22 +460,6 @@ Global
 		{3ECCAF9D-3B23-4980-83E2-8ACEE0FC6BEB}.Release|Any CPU.Build.0 = Debug|x86
 		{3ECCAF9D-3B23-4980-83E2-8ACEE0FC6BEB}.Release|x86.ActiveCfg = Debug|x86
 		{3ECCAF9D-3B23-4980-83E2-8ACEE0FC6BEB}.Release|x86.Build.0 = Debug|x86
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Debug|x86.Build.0 = Debug|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Release|x86.ActiveCfg = Release|Any CPU
-		{BA48309C-FB66-47D3-997D-D3068C4F719B}.Release|x86.Build.0 = Release|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Debug|x86.Build.0 = Debug|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Release|x86.ActiveCfg = Release|Any CPU
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F}.Release|x86.Build.0 = Release|Any CPU
 		{E47E649E-5044-485B-A077-8FB282094A7D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E47E649E-5044-485B-A077-8FB282094A7D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E47E649E-5044-485B-A077-8FB282094A7D}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -582,8 +560,6 @@ Global
 		{A28896EF-67DD-43E1-9AD3-B47270D9DDC5} = {25DE7210-DFC0-448B-894E-84C1C9CA223E}
 		{5DFD3BF7-AA23-4038-AFC6-064E235E37E4} = {E93C2CF9-69A6-4669-BE8A-6060B18FEDCA}
 		{3ECCAF9D-3B23-4980-83E2-8ACEE0FC6BEB} = {5DFD3BF7-AA23-4038-AFC6-064E235E37E4}
-		{BA48309C-FB66-47D3-997D-D3068C4F719B} = {EFA34F81-1751-49D7-8F62-4085E75FA5F2}
-		{B8E86E46-8EC6-4147-AB0D-2C8912F4590F} = {EFA34F81-1751-49D7-8F62-4085E75FA5F2}
 		{E47E649E-5044-485B-A077-8FB282094A7D} = {628A579F-0D27-4F9F-A756-8CC0A67A7B3D}
 		{EE7F939B-7127-433C-80CD-5DB278361229} = {32F0945C-72D0-4116-95D5-F9EE2177D51B}
 		{4F5AE2A8-56DD-44EC-9872-B7884CC1C1DB} = {32F0945C-72D0-4116-95D5-F9EE2177D51B}
