+++ src/Education.UnitTests/Rule/RuleInfoTests.cs
@@ -18,13 +18,9 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System;
-using System.Collections.Generic;
-using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SonarLint.VisualStudio.Core.Analysis;
 
-namespace SonarLint.VisualStudio.Rules.UnitTests
+namespace SonarLint.VisualStudio.Education.UnitTests.Rule
 {
     [TestClass]
     public class RuleInfoTests
