+++ src/Education/Education.csproj
@@ -59,8 +59,6 @@
     <ProjectReference Include="..\Core\Core.csproj" />
     
     <ProjectReference Include="..\Infrastructure.VS\Infrastructure.VS.csproj" />
-    
-    <ProjectReference Include="..\Rules\Rules.csproj" />
   </ItemGroup>
   
   <ItemGroup>
