+++ src/Education/Rule/IRuleInfo.cs
@@ -27,7 +27,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using SonarLint.VisualStudio.Core.Analysis;
 
-namespace SonarLint.VisualStudio.Rules
+namespace SonarLint.VisualStudio.Education.Rule
 {
     public enum RuleIssueSeverity
     {
