+++ src/Education/XamlGenerator/RuleHelpXamlBuilder.cs
@@ -20,7 +20,7 @@
 
 using System.ComponentModel.Composition;
 using System.Windows.Documents;
-using SonarLint.VisualStudio.Rules;
+using SonarLint.VisualStudio.Education.Rule;
 
 namespace SonarLint.VisualStudio.Education.XamlGenerator
 {
