+++ src/Education/XamlGenerator/SubTitleImageInfo.cs
@@ -20,7 +20,7 @@
 
 using System.Collections.Generic;
 using SonarLint.VisualStudio.Core.Analysis;
-using SonarLint.VisualStudio.Rules;
+using SonarLint.VisualStudio.Education.Rule;
 
 namespace SonarLint.VisualStudio.Education.XamlGenerator
 {
