--- src/Rules.UnitTests/Rules.UnitTests.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="..\SonarLint.Test.props" />
-
-  <PropertyGroup>
-    <RootNamespace>SonarLint.VisualStudio.Rules.UnitTests</RootNamespace>
-    <AssemblyName>SonarLint.VisualStudio.Rules.UnitTests</AssemblyName>
-  </PropertyGroup>
-  
-  <ItemGroup Label="Test framework">
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.4.0" />
-
-    <PackageReference Include="Moq" Version="4.18.2" />
-    <PackageReference Include="FluentAssertions" Version="5.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Rules\Rules.csproj" />
-    <ProjectReference Include="..\TestInfrastructure\TestInfrastructure.csproj" />
-  </ItemGroup>
-
-</Project>
