--- src/Rules/Rules.csproj
@@ -1,47 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <!-- NOTE: this assembly should NOT have any dependencies on Visual Studio assemblies -->
-
-  <Import Project="..\SonarLint.Product.props" />
-
-  <PropertyGroup>
-    <RootNamespace>SonarLint.VisualStudio.Rules</RootNamespace>
-    <AssemblyName>SonarLint.VisualStudio.Rules</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup Label="Project references">
-    <!-- Add reference to the ProcessJarFiles project to ensure it is built first
-         (since it unpacks the jar files and copies the .desc files to this project. -->
-    <ProjectReference Include="..\..\build\ProcessJarFiles\ProcessJarFiles.proj">
-      <!-- NB "Private=False" appears as "CopyLocal: No" in the VS properties window -->
-      <Private>False</Private>
-      <ReferenceOutputAssembly>False</ReferenceOutputAssembly>
-    </ProjectReference>
-    <ProjectReference Include="..\Core\Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup Label="Framework references">
-    <Reference Include="System.ComponentModel.Composition" />
-  </ItemGroup>
-
-  <ItemGroup Label="Embedded rule data files">
-    <None Remove="Embedded\**\*.*" />
-    <EmbeddedResource Include="Embedded\**\*.*" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Resources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Update="Resources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-  </ItemGroup>
-
-</Project>
